python-service/requirements.txt
confluent-kafka==2.6.1
fastapi>=0.122.0
uvicorn>=0.38.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-confluent-kafka
opentelemetry-exporter-otlp-proto-grpc
